--- client/components/Add/index.js~master
@@ -1,1 +1,0 @@
-export {default as AddContainer} from './AddContainer'
