--- client/components/AddSuccess/index.js~master
@@ -1,1 +1,0 @@
-export {default as AddSuccess} from './AddSuccess'
