--- client/components/Article/index.js~master
@@ -1,1 +1,0 @@
-export {default as ArticleContainer} from './ArticleContainer'
