--- client/components/Closet/index.js~master
@@ -1,1 +1,0 @@
-export {default as ClosetContainer} from './ClosetContainer'
