--- client/components/Search/index.js~HEAD
@@ -1,13 +1,0 @@
-import React, {Component} from 'react'
-
-class Search extends Component {
-  render() {
-    return (
-      <div>
-        <h1>search</h1>
-      </div>
-    )
-  }
-}
-
-export default Search
