--- client/components/Search/index.js~master
@@ -1,1 +1,0 @@
-export {default as SearchContainer} from './SearchContainer'
