+++ client/routes.js
@@ -34,6 +34,7 @@ class Routes extends Component {
         <Route path="/signup" component={Signup} />
         {isLoggedIn && (
           <Switch>
+            <Route exact path="/" component={Home} />
             <Route path="/home" component={Home} />
             <Route path="/search" component={Search} />
             <Route exact path="/closet" component={Closet} />
